Styling Update to article Item
