+++ client/components/Announcements.jsx
@@ -1,12 +1,11 @@
 /* eslint-disable import/extensions */
 import React, { useState, useEffect } from 'react';
-import axios from 'axios';
 
 import Overlay from './Overlay.jsx';
 
 const Announcements = (props) => {
   const [hover, setHover] = useState(false);
-  const [allGames, setAllGames] = useState([]);
+  const {allGames} = props;
 
   const renderOverlay = () => {
     setHover(true);
@@ -15,21 +14,6 @@ const Announcements = (props) => {
   const unMountOverlay = () => {
     setHover(false);
   };
-
-  const getAllGames = () => {
-    axios.get('/allGames')
-      .then((res) => {
-        setAllGames(res.data);
-      })
-      .catch((err) => {
-        // eslint-disable-next-line no-console
-        console.error(err);
-      });
-  };
-
-  useEffect(() => {
-    getAllGames();
-  }, []); // empty array as dependency required to stop infinite loop
 
   // DOES NOT RENDER UNTIL THE INITIAL GET REQUEST IS COMPLETE
   if (allGames[0] === undefined) {
