+++ client/components/Articles.jsx
@@ -1,8 +1,10 @@
+/* eslint-disable import/extensions */
 /* eslint-disable react/prop-types */
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import ArticleItem from './ArticleItem.jsx';
 
 const Articles = (props) => {
+  const collection = props.allGames[0].announcements;
   return (
     <div id="articles" onClick={props.toggleArticles}>
       <div id="article-item-area">
@@ -11,7 +13,7 @@ const Articles = (props) => {
             <circle r="20" fill="orange" />
           </svg>
         </div>
-        <ArticleItem />
+        <ArticleItem announcement={collection[0]} />
       </div>
     </div>
   );
