+++ public/styles.css
@@ -95,15 +95,16 @@ body {
 }
 
 div#article-item-area {
-  width: 44vw;
+  width: 65vw;
   margin: 0 auto;
 }
 
 div#article-item {
   /* background: #26292e; */
-  height: 34vw;
+  height: 50vw;
   display: grid;
   grid-template-areas:
+      "thumbnail"
       "thumbnail"
       "thumbnail"
       "title"
@@ -116,11 +117,10 @@ div#article-item {
   color: #ffffff;
 }
 
-.article-item-thumbnail {
+.article-item-thumbnail-container {
   display: grid;
   grid-area: thumbnail;
-  grid-row: 1 / span 2;
-  background: green;
+  grid-row: 1 / span 3;
   align-items: center;
 }
 
@@ -129,17 +129,24 @@ div#article-item {
   background: #37393f;
   display: grid;
   grid-area: title;
-  grid-row: 3 / span 1;
+  grid-row: 4 / span 1;
+  text-align: left;
+  padding-left: 3.5em;
+  box-shadow: 0px 0px 5px #1d1e21;
+  z-index: 10;
 }
 
 .article-item-body {
   background: #37393f;
   display: grid;
   grid-area: title;
-  grid-row: 4 / span 3;
+  grid-row: 5 / span 3;
   align-items: baseline;
   text-align: left;
   padding: 2em;
+  font-size: 1.75em;
+  box-shadow: 0px 0px 5px #1d1e21;
+  z-index: 9;
 }
 
 .article-item-interactions {
@@ -148,9 +155,11 @@ div#article-item {
   flex-flow: row;
   grid-area: title;
   justify-content: flex-start;
-  grid-row: 7 / span 2;
+  grid-row: 8 / span 2;
   align-items: center;
   height: 75%;
+  z-index: 11;
+  box-shadow: 0px 0px 5px #1d1e21;
 }
 
 circle {
@@ -290,4 +299,15 @@ img.blue-bubble {
 #share-icon{
   display: inline-flex;
   height: 3em;
+}
+
+span.announcement-title {
+  font-size: 3em;
+}
+
+img.article-item-thumbnail {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  object-position: center;
 }
